EZP-24361: Extracted image variation rendering to dedicated Twig extension
