+++ eZ/Publish/Core/MVC/Symfony/Templating/Twig/Extension/ImageExtension.php
@@ -1,0 +1,80 @@
+<?php
+/**
+ * This file is part of the eZ Publish Kernel package.
+ *
+ * @copyright Copyright (C) eZ Systems AS. All rights reserved.
+ * @license For full copyright and license information view LICENSE file distributed with this source code.
+ * @version //autogentag//
+ */
+
+namespace eZ\Publish\Core\MVC\Symfony\Templating\Twig\Extension;
+
+use eZ\Publish\API\Repository\Exceptions\InvalidVariationException;
+use eZ\Publish\API\Repository\Values\Content\Field;
+use eZ\Publish\API\Repository\Values\Content\VersionInfo;
+use eZ\Publish\Core\MVC\Exception\SourceImageNotFoundException;
+use eZ\Publish\SPI\Variation\VariationHandler;
+use Twig_Extension;
+use Twig_SimpleFunction;
+
+class ImageExtension extends Twig_Extension
+{
+    /**
+     * @var VariationHandler
+     */
+    private $imageVariationService;
+
+    public function __construct( VariationHandler $imageVariationService )
+    {
+        $this->imageVariationService = $imageVariationService;
+    }
+
+    public function getName()
+    {
+        return 'ezpublish.image';
+    }
+
+    public function getFunctions()
+    {
+        return array(
+            new Twig_SimpleFunction(
+                'ez_image_alias',
+                array( $this, 'getImageVariation' ),
+                array( 'is_safe' => array( 'html' ) )
+            ),
+        );
+    }
+
+    /**
+     * Returns the image variation object for $field/$versionInfo
+     *
+     * @param \eZ\Publish\API\Repository\Values\Content\Field $field
+     * @param \eZ\Publish\API\Repository\Values\Content\VersionInfo $versionInfo
+     * @param string $variationName
+     *
+     * @return \eZ\Publish\SPI\Variation\Values\Variation
+     */
+    public function getImageVariation( Field $field, VersionInfo $versionInfo, $variationName )
+    {
+        try
+        {
+            return $this->imageVariationService->getVariation( $field, $versionInfo, $variationName );
+        }
+        catch ( InvalidVariationException $e )
+        {
+            if ( isset( $this->logger ) )
+            {
+                $this->logger->error( "Couldn't get variation '{$variationName}' for image with id {$field->value->id}" );
+            }
+        }
+        catch ( SourceImageNotFoundException $e )
+        {
+            if ( isset( $this->logger ) )
+            {
+                $this->logger->error(
+                    "Couldn't create variation '{$variationName}' for image with id {$field->value->id} because source image can't be found"
+                );
+            }
+        }
+    }
+}
